+ add/remove files to album

+++ src/albums/albums.service.ts
@@ -1,4 +1,10 @@
-import { Inject, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import { FilesService } from 'src/files/files.service';
 import { UsersService } from 'src/users/users.service';
 import { Repository } from 'typeorm';
 import { CreateAlbumDto } from './dto/create-album.dto';
@@ -11,6 +17,7 @@ export class AlbumsService {
     @Inject('ALBUM_REPOSITORY')
     private albumRepository: Repository<Album>,
     private usersService: UsersService,
+    private fileService: FilesService,
   ) {}
 
   async create(createAlbumDto: CreateAlbumDto, userId: number): Promise<Album> {
@@ -40,12 +47,25 @@ export class AlbumsService {
     });
   }
 
+  async findWithFiles(fileId: number): Promise<Album[]> {
+    return await this.albumRepository
+      .createQueryBuilder('album')
+      .leftJoinAndSelect('album.files', 'file')
+      .where('file.id = :fileId', { fileId })
+      .getMany();
+  }
+
   async update(id: number, updateAlbumDto: UpdateAlbumDto): Promise<Album> {
     await this.albumRepository.update(id, updateAlbumDto);
     return this.findOne(id);
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const album = await this.findOne(id);
+    if (album.files.length > 0) {
+      throw new BadRequestException('Album is not empty');
+    }
+
     return this.albumRepository.delete(id);
   }
 
@@ -56,5 +76,34 @@ export class AlbumsService {
     }
 
     return album.user.id === userId;
+  }
+
+  async addFilesToAlbum(
+    album: Album,
+    fileIds: number[],
+    userId: number,
+  ): Promise<void> {
+    const files = await this.fileService.findByIds(fileIds);
+    files.forEach((file) => {
+      if (file.user.id !== userId) {
+        throw new BadRequestException(
+          `File #${file.id} does not belong to user`,
+        );
+      }
+    });
+    album.files = [...album.files, ...files];
+    album.files = album.files.filter(
+      (file, index, self) => self.findIndex((f) => f.id === file.id) === index,
+    );
+    await this.albumRepository.save(album);
+  }
+
+  async removeFilesFromAlbum(
+    album: Album,
+    fileIds: number[],
+    userId: number,
+  ): Promise<void> {
+    album.files = album.files.filter((file) => !fileIds.includes(file.id));
+    await this.albumRepository.save(album);
   }
 }
